Nuked Song and its respective playlist. Replaced all instances with Track.
Tests built on Song are now broken because I don't want to fix them. You can work that out yourself.

The cake is a lie.
The cake is a lie.
The cake is a lie.
The cake is a lie.
The cake is a lie.
The cake is a lie.